refactor: split cli.py into cli/ package (#155)
Split the monolithic 3015-line cli.py into a cli/ package with 11 modules:
- __init__.py: main(), config loading, dispatch (214 lines)
- __main__.py: python -m support (5 lines)
- _common.py: shared helpers (184 lines)
- parsers.py: all argparse definitions (490 lines)
- compare.py: logprobs, output, streaming comparison (193 lines)
- batch.py: batch-compare, rerun, snapshot replay (418 lines)
- diagnose.py: diagnose, healthcheck, detect, check-kv (83 lines)
- report.py: report, diff, regression, summary, filter, explain, annotate, cluster, aggregate (338 lines)
- benchmark.py: benchmark, reproducibility, concurrency-sweep, bisect (163 lines)
- data.py: snapshot capture, cache, history, dataset-stats (368 lines)
- config_cmd.py: init, config, profiles, completion (93 lines)
- analysis.py: watch, entropy, length-bias, sensitivity, fingerprint (217 lines)

All 1064 existing tests pass. No functional changes.
