+++ src/xpyd_acc/cli/__main__.py
@@ -1,0 +1,5 @@
+"""Allow ``python -m xpyd_acc.cli``."""
+
+from xpyd_acc.cli import main
+
+main()
